pyproject.toml
[tool.poetry]
name = "vertexcover-assignment"
version = "0.1.0"
description = ""
authors = ["hungerarray <ranjan.ojha01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyaudio = "^0.2.14"
asyncio = "^3.4.3"
webrtcvad = "^2.0.10"
setuptools = "^75.2.0"
deepgram-sdk = "^3.7.5"
openai = "^1.52.0"
python-dotenv = "^1.0.1"

[tool.pyright]
executionEnvironments = [{ root = "vertexcover_assignment/", extraPaths = ["vertexcover_assignment"] }]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
